fix travis
