+++ js/render-block.js
@@ -1,5 +1,5 @@
-const main = document.querySelector(`main`);
 const renderBlock = (blockElement) => {
+  const main = document.querySelector(`main`);
   main.innerHTML = ``;
   main.appendChild(blockElement);
 };
